AppSetting File and Lock added

+++ PersonApp/Repository/PersonRepository.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Routing.Constraints;
+using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
 using Newtonsoft.Json;
 using PersonApp.Model;
@@ -7,7 +8,9 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Reflection.Metadata.Ecma335;
+using System.Threading;
 using System.Threading.Tasks;
+using System.Xml.Linq;
 
 namespace PersonApp.Repository
 {
@@ -15,37 +18,47 @@ namespace PersonApp.Repository
     {
         private readonly IWebHostEnvironment _env;
         private readonly ILogger<PersonRepository> _logger;
+        private readonly IConfiguration _configuration;
+        private readonly object _fileAccess = new object();
 
-        public string FileName => "persons.json"; 
 
-        public PersonRepository(IWebHostEnvironment env, ILogger<PersonRepository> logger)
+        private string JsonFilePath => "JsonFilePath"; 
+
+        public PersonRepository(IWebHostEnvironment env, ILogger<PersonRepository> logger, IConfiguration configuration)
         {
             _env = env;
             _logger = logger;
+            _configuration = configuration;
         }
-
+            
 
         public async Task<bool> Save(Person person)
         {
             try
-            {
+            { 
+
                 List<Person> personlist = new List<Person>();
+               
+                var path = _configuration[JsonFilePath];  
 
-                var path = Path.Combine(_env.ContentRootPath, FileName);
-
-                using (StreamReader r = new StreamReader(path))
+                lock (_fileAccess)
                 {
-                    string json = r.ReadToEnd();
-                    personlist = JsonConvert.DeserializeObject<List<Person>>(json);
+                   if (File.Exists(path))
+                    {
+                        using (StreamReader r = new StreamReader(path))
+                        {
+                            string json = r.ReadToEnd();
+                            personlist = JsonConvert.DeserializeObject<List<Person>>(json);
 
+                        }
+                    }
+                   
+                    personlist.Add(person);
+                    string jsonToWrite = JsonConvert.SerializeObject(personlist.ToArray(), Formatting.Indented);
+                    File.WriteAllText(path, jsonToWrite);
+              
                 }
 
-                personlist.Add(person);
-
-
-                string jsonToWrite = JsonConvert.SerializeObject(personlist.ToArray(), Formatting.Indented);
-
-                File.WriteAllText(path, jsonToWrite);
                 return await Task.FromResult(true);
             }
             catch (Exception ex)
@@ -61,20 +74,27 @@ namespace PersonApp.Repository
 
         public async Task<IEnumerable<Person>> GetAll()
         {
-            var path = Path.Combine(_env.ContentRootPath, FileName);
+            IEnumerable<Person> personlist = null;
+ 
+            var path = _configuration[JsonFilePath];
 
-            using (StreamReader r = new StreamReader(path))
+            if (File.Exists(path))
             {
-                string json = r.ReadToEnd();
-                var _data = JsonConvert.DeserializeObject<IEnumerable<Person>>(json);
-                return await Task.FromResult(_data);
+                lock (_fileAccess)
+                {
+                    using (StreamReader r = new StreamReader(path))
+                    {
+                        string json = r.ReadToEnd();
+                        personlist = JsonConvert.DeserializeObject<IEnumerable<Person>>(json);
+                    }
 
+                }
             }
-
-
+            return await Task.FromResult(personlist);
 
         }
 
+      
     }
 
 }
